+++ src/QueryCat.Backend/Parser/AstBuilder.cs
@@ -10,8 +10,6 @@ namespace QueryCat.Backend.Parser;
 /// </summary>
 internal static class AstBuilder
 {
-    private static readonly ProgramAntlrErrorListener ErrorListener = new();
-
     public static ProgramNode BuildProgramFromString(string program)
         => Build<ProgramNode>(program, p => p.program());
 
@@ -20,17 +18,19 @@ internal static class AstBuilder
 
     private static TNode Build<TNode>(string input, Func<QueryCatParser, ParserRuleContext> signatureFunc) where TNode : IAstNode
     {
+        var errorListener = new ProgramAntlrErrorListener();
+
         var inputStream = new AntlrInputStream(input);
         var lexer = new QueryCatLexer(inputStream);
         var commonTokenStream = new CommonTokenStream(lexer);
         var parser = new QueryCatParser(commonTokenStream);
         parser.RemoveErrorListeners();
-        parser.AddErrorListener(ErrorListener);
+        parser.AddErrorListener(errorListener);
         var context = signatureFunc.Invoke(parser);
         var visitor = new ProgramParserVisitor();
         if (parser.NumberOfSyntaxErrors > 0)
         {
-            throw new SyntaxException("Syntax error.");
+            throw new SyntaxException(errorListener.Message, input, errorListener.Line, errorListener.CharPosition);
         }
         return (TNode)visitor.Visit(context);
     }